requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0

# Mamba SSM (optional - will use pure PyTorch fallback if not available)
mamba-ssm>=1.0.0
causal-conv1d>=1.0.0

# Experiment Tracking
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Profiling & Analysis
fvcore>=0.1.5
thop>=0.1.1

# Optional: Weights & Biases
wandb>=0.13.0

# Scikit-learn for metrics
scikit-learn>=1.0.0

# Einops for tensor operations
einops>=0.6.0
